+++ microservice/microservice/Api/Stats/MicroserviceStatsApi.cs
@@ -38,6 +38,10 @@ namespace Beamable.Server.Api.Stats
       {
          return GetStats("game", "private", "player", userId, stats);
       }
+      public Promise<Dictionary<string, string>> GetAllProtectedPlayerStats(long userId)
+      {
+         return GetStats("game", "private", "player", userId, null);
+      }
 
       public Promise<EmptyResponse> SetProtectedPlayerStat(long userId, string key, string value)
       {
@@ -62,8 +66,8 @@ namespace Beamable.Server.Api.Stats
          string[] stats)
       {
          var key = $"{domain}.{access}.{type}.{userId}";
-         var statString = string.Join(",", stats);
-         return Requester.Request<StatsResponse>(Method.GET, $"{OBJECT_SERVICE}/{key}", new
+         var statString = stats == null ? string.Empty : string.Join(",", stats);
+         return Requester.Request<StatsResponse>(Method.GET, $"{OBJECT_SERVICE}/{key}", new StatsRequest
          {
             stats = statString
          }).Map(res =>
@@ -110,6 +114,11 @@ namespace Beamable.Server.Api.Stats
 
       public long id;
       public Dictionary<string, object> stats;
+   }
+
+   class StatsRequest
+   {
+      public string stats;
    }