+++ client/Packages/com.beamable.server/CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unpublished]
 ### Added
 - Folding/Unfolding services cards in `Microservice Manager`
+- Add `IMicroserviceStatsApi.GetAllProtectedPlayerStats` method without filtering stats
+
+### Fixed
+- Filters now work for `IMicroserviceStatsApi.GetProtectedPlayerStats`
 
 ## [0.17.2]
 ### Fixed